refactor: delete commented code, rename func
