+++ src/index.tsx
@@ -3,21 +3,21 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
+import { apiService } from './services/api-request';
 import App from './components/app/app';
 import toolkitSlice from './reducer';
-import { myApi } from './services/apiRequest';
 
 import './index.css';
 
 const rootReducer = combineReducers({
   toolkit: toolkitSlice,
-  [myApi.reducerPath]: myApi.reducer,
+  [apiService.reducerPath]: apiService.reducer,
 });
 
 export const store = configureStore({
   reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(myApi.middleware),
+    getDefaultMiddleware().concat(apiService.middleware),
 });
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
